change from double to float.

+++ BitbloqCompass/BitbloqCompass.h
@@ -77,45 +77,45 @@
 /* Exported macro ------------------------------------------------------------*/
 //#define COMPASS_SERIAL_DEBUG
 
-#define I2C_ERROR                  (-1)
+#define I2C_ERROR (-1)
 
 // Me Compass only has one address
-#define COMPASS_DEFAULT_ADDRESS    (0x1E)   
+#define COMPASS_DEFAULT_ADDRESS (0x1E)
 
 //Me Compass Register Address
-#define COMPASS_RA_CONFIG_A        (0x00)
-#define COMPASS_RA_CONFIG_B        (0x01)
-#define COMPASS_RA_MODE            (0x02)
-#define COMPASS_RA_DATAX_H         (0x03)
-#define COMPASS_RA_DATAX_L         (0x04)
-#define COMPASS_RA_DATAZ_H         (0x05)
-#define COMPASS_RA_DATAZ_L         (0x06)
-#define COMPASS_RA_DATAY_H         (0x07)
-#define COMPASS_RA_DATAY_L         (0x08)
-#define COMPASS_RA_STATUS          (0x09)
-#define COMPASS_RA_ID_A            (0x0A)
-#define COMPASS_RA_ID_B            (0x0B)
-#define COMPASS_RA_ID_C            (0x0C)
+#define COMPASS_RA_CONFIG_A (0x00)
+#define COMPASS_RA_CONFIG_B (0x01)
+#define COMPASS_RA_MODE (0x02)
+#define COMPASS_RA_DATAX_H (0x03)
+#define COMPASS_RA_DATAX_L (0x04)
+#define COMPASS_RA_DATAZ_H (0x05)
+#define COMPASS_RA_DATAZ_L (0x06)
+#define COMPASS_RA_DATAY_H (0x07)
+#define COMPASS_RA_DATAY_L (0x08)
+#define COMPASS_RA_STATUS (0x09)
+#define COMPASS_RA_ID_A (0x0A)
+#define COMPASS_RA_ID_B (0x0B)
+#define COMPASS_RA_ID_C (0x0C)
 
 //define number of samples averaged per measurement
-#define COMPASS_AVERAGING_1        (0x00)
-#define COMPASS_AVERAGING_2        (0x20)
-#define COMPASS_AVERAGING_4        (0x40)
-#define COMPASS_AVERAGING_8        (0x60)
+#define COMPASS_AVERAGING_1 (0x00)
+#define COMPASS_AVERAGING_2 (0x20)
+#define COMPASS_AVERAGING_4 (0x40)
+#define COMPASS_AVERAGING_8 (0x60)
 
 //define data output rate value (Hz)
-#define COMPASS_RATE_0P75          (0x00)   // 0.75 (Hz)
-#define COMPASS_RATE_1P5           (0x40)   // 1.5  (Hz)
-#define COMPASS_RATE_3             (0x08)   // 3    (Hz)
-#define COMPASS_RATE_7P5           (0x0C)   // 7.5  (Hz)
-#define COMPASS_RATE_15            (0x10)   // 15   (Hz)
-#define COMPASS_RATE_30            (0x14)   // 30   (Hz)
-#define COMPASS_RATE_75            (0x18)   // 75   (Hz)
+#define COMPASS_RATE_0P75 (0x00) // 0.75 (Hz)
+#define COMPASS_RATE_1P5 (0x40)  // 1.5  (Hz)
+#define COMPASS_RATE_3 (0x08)    // 3    (Hz)
+#define COMPASS_RATE_7P5 (0x0C)  // 7.5  (Hz)
+#define COMPASS_RATE_15 (0x10)   // 15   (Hz)
+#define COMPASS_RATE_30 (0x14)   // 30   (Hz)
+#define COMPASS_RATE_75 (0x18)   // 75   (Hz)
 
 //define measurement bias value
-#define COMPASS_BIAS_NORMAL        (0x00)
-#define COMPASS_BIAS_POSITIVE      (0x01)
-#define COMPASS_BIAS_NEGATIVE      (0x02)
+#define COMPASS_BIAS_NORMAL (0x00)
+#define COMPASS_BIAS_POSITIVE (0x01)
+#define COMPASS_BIAS_NEGATIVE (0x02)
 
 //define magnetic field gain value
 /* -+-------------+-----------------
@@ -130,26 +130,26 @@
  *  | +/- 5.6 Ga  | 330
  *  | +/- 8.1 Ga  | 230
  * -+-------------+-----------------*/
-#define COMPASS_GAIN_1370          (0x00)
-#define COMPASS_GAIN_1090          (0x20)
-#define COMPASS_GAIN_820           (0x40)
-#define COMPASS_GAIN_660           (0x60)
-#define COMPASS_GAIN_440           (0x80)
-#define COMPASS_GAIN_390           (0xA0)
-#define COMPASS_GAIN_330           (0xC0)
-#define COMPASS_GAIN_220           (0xE0)
+#define COMPASS_GAIN_1370 (0x00)
+#define COMPASS_GAIN_1090 (0x20)
+#define COMPASS_GAIN_820 (0x40)
+#define COMPASS_GAIN_660 (0x60)
+#define COMPASS_GAIN_440 (0x80)
+#define COMPASS_GAIN_390 (0xA0)
+#define COMPASS_GAIN_330 (0xC0)
+#define COMPASS_GAIN_220 (0xE0)
 
 //define measurement mode
-#define COMPASS_MODE_CONTINUOUS    (0x00)
-#define COMPASS_MODE_SINGLE        (0x01)
-#define COMPASS_MODE_IDLE          (0x02)
+#define COMPASS_MODE_CONTINUOUS (0x00)
+#define COMPASS_MODE_SINGLE (0x01)
+#define COMPASS_MODE_IDLE (0x02)
 
 //define others parameter
 #define COMPASS_PI 3.14159265F
-#define START_ADDRESS_OF_EEPROM_BUFFER  (int16_t)(0x00)
+#define START_ADDRESS_OF_EEPROM_BUFFER (int16_t)(0x00)
 
-
-namespace Bitbloq{
+namespace Bitbloq
+{
 
 /* define a struct to save calibration parameters------------------------------*/
 struct Compass_Calibration_Parameter
@@ -160,7 +160,7 @@ struct Compass_Calibration_Parameter
   float X_gain;
   float Y_gain;
   float Z_gain;
-  uint8_t Rotation_Axis;   //1:X_Axis   2:Y_Axis   3:Z_Axis
+  uint8_t Rotation_Axis; //1:X_Axis   2:Y_Axis   3:Z_Axis
 
   uint8_t verify_flag;
 };
@@ -174,8 +174,7 @@ struct Compass_Calibration_Parameter
 class Compass
 {
 public:
-
-/**
+  /**
  * Alternate Constructor which can call your own function to map the _keyPin and _ledPin to arduino port,
  * no pins are used or initialized here
  * \param[in]
@@ -185,7 +184,7 @@ public:
  */
   Compass(uint8_t keyPin, uint8_t ledPin);
 
-/**
+  /**
  * Alternate Constructor which can call your own function to map the _keyPin and _ledPin to arduino port
  * and change the i2c device address, no pins are used or initialized here
  * \param[in]
@@ -197,7 +196,7 @@ public:
  */
   Compass(uint8_t keyPin, uint8_t ledPin, uint8_t address);
 
-/**
+  /**
  * \par Function
  *   setpin
  * \par Description
@@ -215,7 +214,7 @@ public:
  */
   void setpin(uint8_t keyPin, uint8_t ledPin);
 
-/**
+  /**
  * \par Function
  *   setup
  * \par Description
@@ -228,10 +227,10 @@ public:
  *   None
  * \par Others
  *   You can check the HMC5883 datasheet for the macro definition.
- */    
+ */
   void setup(void);
 
-/**
+  /**
  * \par Function
  *   testConnection
  * \par Description
@@ -247,7 +246,7 @@ public:
  */
   bool testConnection(void);
 
-/**
+  /**
  * \par Function
  *   getAngle
  * \par Description
@@ -261,9 +260,9 @@ public:
  * \par Others
  *   Will return a correct angle when you keep the Compass working in the plane which have calibrated.
  */
-  double getAngle(void);  
+  float getAngle(void);
 
-/**
+  /**
  * \par Function
  *   getHeadingX
  * \par Description
@@ -279,7 +278,7 @@ public:
  */
   int16_t getHeadingX(void);
 
-/**
+  /**
  * \par Function
  *   getHeadingY
  * \par Description
@@ -295,7 +294,7 @@ public:
  */
   int16_t getHeadingY(void);
 
-/**
+  /**
  * \par Function
  *   getHeadingZ
  * \par Description
@@ -311,7 +310,7 @@ public:
  */
   int16_t getHeadingZ(void);
 
-/**
+  /**
  * \par Function
  *   getHeading
  * \par Description
@@ -330,16 +329,17 @@ public:
  *   The sequence of the sensor data registors of HMC5883 is X, Z, Y.
  */
   int16_t getHeading(int16_t *x, int16_t *y, int16_t *z);
+
 private:
-  static volatile uint8_t  _keyPin;
-  static volatile uint8_t  _ledPin;
+  static volatile uint8_t _keyPin;
+  static volatile uint8_t _ledPin;
   bool Calibration_Flag;
   uint8_t buffer[6];
   uint8_t Device_Address;
   uint8_t Measurement_Mode;
   struct Compass_Calibration_Parameter Cal_parameter;
 
-/**
+  /**
  * \par Function
  *   writeReg
  * \par Description
@@ -361,10 +361,10 @@ private:
  *   refer to the arduino official library twi.c
  * \par Others
  *   To set the registor for initializing.
- */ 
+ */
   int8_t writeReg(int16_t reg, uint8_t data);
 
-/**
+  /**
  * \par Function
  *   writeData
  * \par Description
